Fix signedness mp3 stream compiler warning.

+++ src/mp3inputstream.hh
@@ -16,7 +16,7 @@ class MP3InputStream : public AudioInputStream
   int         m_n_values = 0;
   int         m_n_channels = 0;
   int         m_sample_rate = 0;
-  int         m_frames_left = 0;
+  size_t      m_frames_left = 0;
   bool        m_need_close = false;
   bool        m_eof = false;
   State       m_state = State::NEW;
